requirements.txt
# ==========================================
# 1. 核心 Web 框架与服务器
# ==========================================
fastapi==0.125.0        # 高性能 Web 框架，用于构建后端 API
uvicorn==0.38.0         # ASGI 服务器，用于运行 FastAPI
python-multipart==0.0.21 # 必装！用于处理文件上传 (UploadFile)
httpx==0.28.1           # 异步 HTTP 客户端，LangChain 底层依赖它调用 API
requests==2.32.5        # 同步 HTTP 客户端，用于测试脚本

# ==========================================
# 2. 大模型与编排 (LangChain 全家桶)
# ==========================================
langchain==1.2.0            # LangChain 主框架
langchain-community==0.4.1  # 社区组件 (包含各种 Loader, VectorStore)
langchain-core==1.2.2       # 核心接口定义
langchain-openai==1.1.5     # 用于调用 DeepSeek (兼容 OpenAI 协议)
langchain-huggingface==1.2.0 # 用于加载本地 Embedding 模型
langchain-text-splitters==1.1.0 # 用于文档递归切分
langsmith==0.5.0            # 用于链路追踪和可视化调试

# ==========================================
# 3. RAG 核心 (向量库与模型)
# ==========================================
faiss-cpu==1.13.2           # 向量数据库 (Facebook开源)，CPU版，稳定且轻量
sentence-transformers==5.2.0 # 用于运行 BGE-Reranker 和 Embedding 模型
torch==2.9.1                # PyTorch 深度学习框架 (模型底层依赖)
numpy==2.3.5                # 数学运算库
scikit-learn==1.8.0         # 机器学习工具库

# ==========================================
# 4. 文档处理
# ==========================================
pypdf==6.5.0                # 用于解析 PDF 文件内容
docling>=2.0.0              # 结构化 PDF 解析（用于 Docling 模式）

# ==========================================
# 5. 数据库与 ORM
# ==========================================
SQLAlchemy==2.0.45          # ORM 框架，用于操作 SQLite 数据库
PyMySQL==1.1.2              # MySQL 驱动
redis==7.0.1                # 缓存与轻量任务协同
alembic==1.14.1             # 数据库迁移

# ==========================================
# 6. 安全与认证
# ==========================================
passlib==1.7.4              # 密码哈希工具
bcrypt==5.0.0               # 加密算法支持
python-dotenv==1.2.1        # 用于读取 .env 环境变量

# ==========================================
# 7. 底层支撑依赖 (自动安装，建议保留以锁定版本)
# ==========================================
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
marshmallow==3.26.1
multidict==6.7.0
openai==2.13.0
orjson==3.11.5
packaging==25.0
protobuf==6.33.2
pydantic==2.12.5
pydantic-core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
safetensors==0.7.0
scipy==1.16.3
six==1.17.0
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.3.0
watchfiles==1.1.1
yarl==1.22.0
zipp==3.23.0
# chromadb==1.3.7  # (注：已切换至 FAISS，保留此行以防旧代码依赖，如确认不用可删除)
